+++ spec/features/admin/buildings/create_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.feature 'Building create' do
+  before do
+    log_in create(:user, role: 'superuser')
+    visit admin_buildings_path
+  end
+
+  it 'succeeds' do
+    click_on 'New building'
+    fill_in 'Name', with: 'Test'
+    click_on 'Create'
+    expect(page).to have_content('Building was successfully created.')
+  end
+end